requirements.txt
pick
pillow
numpy